IG-360 - Added sort functionality Cleaned up search and filter backend
